+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/verify/IdleVerifyUtility.java
@@ -46,7 +46,6 @@ import org.apache.ignite.internal.processors.cache.distributed.dht.topology.Grid
 import org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtPartitionTopology;
 import org.apache.ignite.internal.processors.cache.persistence.CacheDataRow;
 import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStore;
-import org.apache.ignite.internal.processors.cache.persistence.snapshot.IncrementalSnapshotVerificationTask.HashHolder;
 import org.apache.ignite.internal.util.lang.GridIterator;
 import org.apache.ignite.internal.util.lang.IgniteThrowableSupplier;
 import org.apache.ignite.internal.util.typedef.F;
@@ -381,11 +380,12 @@ public class IdleVerifyUtility {
         }
 
         /**
-         * @param hash Incremental snapshot hash holder.
+         * @param partHash Partition hash.
+         * @param partHash Version hash.
          */
-        public VerifyPartitionContext(HashHolder hash) {
-            this.partHash = hash.hash;
-            this.partVerHash = hash.verHash;
+        public VerifyPartitionContext(int partHash, int partVerHash) {
+            this.partHash = partHash;
+            this.partVerHash = partVerHash;
         }
 
         /** */
